notionposter_windows/requirements.txt
py-notifier==0.3.2
Requests==2.31.0
flox-lib==0.20.1